# PERF: Optimize Vector.cs