Hi @CreoOne, I've tried to make some performance optimizations to this repo. Following are the benchmark results before and after the changes:

**Benchmarks before changes:**
| Method | Dimensionality |            Mean |         Error |         StdDev |        Gen 0 |    Allocated |
|------- |--------------- |----------------:|--------------:|---------------:|-------------:|-------------:|
|  Cross |              3 |        19.17 us |      2.172 us |       6.090 us |            - |         4 KB |
|  Cross |             10 | 1,553,068.10 us | 53,050.037 us | 144,326.538 us | 1232000.0000 | 2,517,116 KB |

**Benchmarks after changes:**
| Method | Dimensionality |            Mean |         Error |        StdDev |          Median |        Gen 0 |    Allocated |
|------- |--------------- |----------------:|--------------:|--------------:|----------------:|-------------:|-------------:|
|  Cross |              3 |        14.75 us |      2.019 us |      5.824 us |        13.10 us |            - |         4 KB |
|  Cross |             10 | 1,265,088.71 us | 23,984.376 us | 22,435.000 us | 1,266,821.10 us | 1172000.0000 | 2,394,936 KB |

As you can see the allocations have gone down. I've also made sure that all the unit tests pass following these changes as well. Could you please confirm whether or not you think my changes are valid?

Thank you!